functions/inference/requirements.txt
# Function dependencies, for example:
# package>=version

numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
google-cloud-storage==1.19.0
tensorflow==1.13.1